Movie recommendation program + movies descriptions
This code provides a movie recommendation system based on the word vector similarity of the description of movies (spyCy). The function takes in the description as a parameter and returns the most similar movie.
--------------------------------------------------------
In the provided file 'movies.txt' there are no movie titles, only 'Movie A', 'Movie B' etc. Therefore, I treat these as movie titles, and for clarity I assign a provided description to the chosen title 'Movie x' as an extra output.